pyproject.toml
[project]
name = "nuclea-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-mcp-adapters>=0.1.4",
    "langchain-openai>=0.3.19",
    "langgraph>=0.4.8",
    "mcp[cli]>=1.9.2",
    "python-dotenv>=1.1.0",
]
